cap nhat fixbug

+++ web/src/main/java/com/cmms/webapp/action/MachineAction.java
@@ -9,18 +9,25 @@ import com.cmms.model.Machine;
 import com.cmms.model.MachineType;
 import com.cmms.util.DateUtil;
 import com.cmms.webapp.security.LoginSuccessHandler;
+import com.cmms.webapp.util.ImageUtil;
 import com.cmms.webapp.util.ResourceBundleUtils;
 import com.cmms.webapp.util.WebUtil;
 import com.opensymphony.xwork2.Preparable;
 import java.io.ByteArrayInputStream;
+import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileNotFoundException;
+import java.io.FileOutputStream;
+import java.io.IOException;
 import java.io.InputStream;
+import java.io.OutputStream;
 import java.sql.Timestamp;
 import java.util.ArrayList;
-import java.util.Calendar;
 import java.util.Date;
 import java.util.List;
 import java.util.Map;
 import org.apache.commons.lang.StringUtils;
+import org.apache.struts2.ServletActionContext;
 import org.codehaus.jettison.json.JSONArray;
 import org.codehaus.jettison.json.JSONObject;
 
@@ -29,6 +36,7 @@ import org.codehaus.jettison.json.JSONObject;
  */
 public class MachineAction extends BaseAction implements Preparable {
 
+    //<editor-fold defaultstate="collapsed" desc="Other">
     private static final long serialVersionUID = -1L;
     private ItemTypeDao itemTypeDao;
     private MachineDao machineDao;
@@ -73,8 +81,9 @@ public class MachineAction extends BaseAction implements Preparable {
 
     public String index() {
         return SUCCESS;
-    }
+    }//</editor-fold>
 
+    //<editor-fold defaultstate="collapsed" desc="getTree">
     public InputStream getTree() {
         try {
             String idReq = getRequest().getParameter("id");
@@ -109,8 +118,9 @@ public class MachineAction extends BaseAction implements Preparable {
             log.error("ERROR getTree: ", e);
             return null;
         }
-    }
+    }//</editor-fold>
 
+    //<editor-fold defaultstate="collapsed" desc="getListSytem2">
     private List<Integer> getListSytem2() {
         List<Integer> listCompany = getListSytem();
         if (listCompany == null) {
@@ -120,8 +130,9 @@ public class MachineAction extends BaseAction implements Preparable {
             }
         }
         return listCompany;
-    }
+    }//</editor-fold>
 
+    //<editor-fold defaultstate="collapsed" desc="getLoadData">
     public InputStream getLoadData() {
         try {
             JSONObject result = new JSONObject();
@@ -211,8 +222,9 @@ public class MachineAction extends BaseAction implements Preparable {
             log.error("ERROR getLoadData: ", ex);
             return null;
         }
-    }
+    }//</editor-fold>
 
+    //<editor-fold defaultstate="collapsed" desc="getSave">
     public InputStream getSave() {
         try {
             JSONObject result = new JSONObject();
@@ -229,6 +241,8 @@ public class MachineAction extends BaseAction implements Preparable {
             String note = getRequest().getParameter("note");
             String since = getRequest().getParameter("since");
             String completeCode = getRequest().getParameter("completeCode");
+            String imgUrl = getRequest().getParameter("imgUrl");
+            String imgPath = getRequest().getParameter("imgPath");
 
             if (StringUtils.isBlank(code)) {
                 result.put("success", false);
@@ -286,7 +300,8 @@ public class MachineAction extends BaseAction implements Preparable {
             machenic.setSpecification(specification);
             machenic.setNote(note);
             machenic.setCompleteCode(completeCode);
-            
+            machenic.setImgPath(imgPath);
+            machenic.setImgUrl(imgUrl);
             machenic = machineDao.save(machenic);
             if (machenic != null) {
                 result.put("success", true);
@@ -300,8 +315,9 @@ public class MachineAction extends BaseAction implements Preparable {
             log.error("ERROR getSave: ", ex);
             return null;
         }
-    }
+    }//</editor-fold>
 
+    //<editor-fold defaultstate="collapsed" desc="getDelete">
     private String[] ids;
 
     public String[] getIds() {
@@ -324,7 +340,14 @@ public class MachineAction extends BaseAction implements Preparable {
                         result.put("message", ResourceBundleUtils.getName("deleteUsing"));
                         return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
                     }
+                    if (machineDao.checkUseWO(list)) {
+                        result.put("success", false);
+                        result.put("message", ResourceBundleUtils.getName("deleteUsing"));
+                        return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
+                    }
+                    String path = machineDao.get(Long.parseLong(ids[0])).getImgPath();
                     machineDao.remove(Long.parseLong(ids[0]));
+                    WebUtil.deleteFile(path);
                 } else {
                     for (String idTmp : ids) {
                         list.add(Long.parseLong(idTmp));
@@ -334,12 +357,26 @@ public class MachineAction extends BaseAction implements Preparable {
                         result.put("message", ResourceBundleUtils.getName("deleteUsing"));
                         return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
                     }
+                    if (machineDao.checkUseWO(list)) {
+                        result.put("success", false);
+                        result.put("message", ResourceBundleUtils.getName("deleteUsing"));
+                        return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
+                    }
+                    String[] lstPath = new String[ids.length];
+                    int i = 0;
+                    for (String idTmp : ids) {
+                        lstPath[i++] = machineDao.get(Long.parseLong(idTmp)).getImgPath();
+                    }
                     int delete = machineDao.delete(list);
                     if (delete != ids.length) {
+
                         log.warn("deleteCompany rtn: " + delete + " list: " + ids.length);
                         result.put("success", false);
                         result.put("message", ResourceBundleUtils.getName("deleteFail"));
                         return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
+                    }
+                    for (String path : lstPath) {
+                        WebUtil.deleteFile(path);
                     }
                 }
                 result.put("success", true);
@@ -353,5 +390,100 @@ public class MachineAction extends BaseAction implements Preparable {
             log.error("ERROR getDelete: ", ex);
             return null;
         }
+    }//</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="getSaveImg">
+    public InputStream getSaveImg() {
+        try {
+            JSONObject result = new JSONObject();
+            String rtn = saveFile();
+            if (rtn != null) {
+                String uploadDir = ServletActionContext.getServletContext().getRealPath("/");
+                uploadDir += File.separator + PATH_UPLOAD + File.separator + PATH_MECHANIC + File.separator;
+                result.put("success", true);
+                result.put("url", "../" + PATH_UPLOAD + "/" + PATH_MECHANIC + "/" + rtn);
+                result.put("path", uploadDir + rtn);
+                result.put("message", ResourceBundleUtils.getName("uploadSuccessMsg"));
+            } else {
+                result.put("success", false);
+                result.put("message", ResourceBundleUtils.getName("uploadFailMsg"));
+            }
+            return new ByteArrayInputStream(result.toString().getBytes("UTF8"));
+        } catch (Exception ex) {
+            log.error("ERROR getSaveImg: ", ex);
+            return null;
+        }
+    }//</editor-fold>
+
+    //<editor-fold defaultstate="collapsed" desc="saveFile">
+    private File file;
+    private String fileFileName;
+
+    public File getFile() {
+        return file;
     }
+
+    public void setFile(File file) {
+        this.file = file;
+    }
+
+    public String getFileFileName() {
+        return fileFileName;
+    }
+
+    public void setFileFileName(String fileFileName) {
+        this.fileFileName = fileFileName;
+    }
+
+    private String saveFile() {
+        InputStream stream = null;
+        String rtn = null;
+        try {
+            String fileName = WebUtil.removeAscii(fileFileName);
+            fileName = fileName.replaceAll("\\s+", "_");
+            String uploadDir = ServletActionContext.getServletContext().getRealPath("/");
+            uploadDir += File.separator + PATH_UPLOAD + File.separator + PATH_MECHANIC + File.separator;
+
+            File theDir = new File(uploadDir);
+            // if the directory does not exist, create it
+            if (!theDir.exists()) {
+                log.info("mkdirs: " + uploadDir);
+                try {
+                    theDir.mkdirs();
+                } catch (Exception ex) {
+                    //handle it
+                    log.info("mkdirs Exception: ", ex);
+                }
+            }
+
+            stream = new FileInputStream(this.file);
+            String fullPath = uploadDir + "tmp_" + fileName;
+            fullPath = fullPath.replace("\\", File.separator);
+            // write the file to the file specified
+            log.info("-------fullPath: " + fullPath);
+            try (OutputStream bos = new FileOutputStream(fullPath)) {
+                int bytesRead;
+                byte[] buffer = new byte[8192];
+                while ((bytesRead = stream.read(buffer, 0, 8192)) != -1) {
+                    bos.write(buffer, 0, bytesRead);
+                }
+                bos.flush();
+            }
+            ImageUtil.resize(fullPath, uploadDir + fileName);
+            rtn = fileName;
+        } catch (FileNotFoundException ex) {
+            log.error("ERROR FileNotFoundException: ", ex);
+        } catch (Exception ex) {
+            log.error("ERROR Exception: ", ex);
+        } finally {
+            try {
+                if (stream != null) {
+                    stream.close();
+                }
+            } catch (IOException ex) {
+                log.error("ERROR IOException: ", ex);
+            }
+        }
+        return rtn;
+    }//</editor-fold>
 }
